requirements.txt
Django==1.8.2
Pillow==2.8.2
argparse==1.2.1
djangorestframework==3.1.3
wsgiref==0.1.2
